TUP-16584：Job run Tab >> Advanced Settings >> Add JVM argument - No escape character works [MacOS]
